hw3 done

+++ src/main/java/com/example/taxiservicespring/controller/dto/CarDto.java
@@ -10,6 +10,6 @@ public class CarDto {
     private String regNumber;
     private long modelId;
     private int categoryId;
-    private long locationId;
+    private int locationId;
     private CarStatus status;
 }
